requirements.txt
streamlit
pandas
pdfplumber
pillow
pytesseract
python-dotenv
groq
opencv-python-headless